# Punycode speed improvements