I had this hanging around on a branch, and I figured I should pull it forward now rather than let linger. A rough benchmark using the input `"http://日本.co.jp" "/page"`:

```
Old:
Benchmarking parse + punycode with 1000000 per run:
  Average: 1043.73 ms
     Rate: 958.101 k-iter / s
Benchmarking punycode + unpunycode with 1000000 per run:
  Average: 1566.77 ms
     Rate: 638.256 k-iter / s

New:
Benchmarking parse + punycode with 1000000 per run:
  Average: 727.279 ms
     Rate: 1374.99 k-iter / s
Benchmarking punycode + unpunycode with 1000000 per run:
  Average: 946.035 ms
     Rate: 1057.04 k-iter / s
```

@b4hand @tammybailey @lindseyreno